Add Resume button
Also break down config
