+++ src/pages/index.tsx
@@ -1,21 +1,29 @@
 import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
 import Layout from "@theme/Layout";
 import clsx from "clsx";
-import React from "react";
+import React, { useMemo } from "react";
 import styles from "./index.module.css";
 
 function HomepageHeader() {
   const { siteConfig } = useDocusaurusContext();
+  const lastResumeUpdate = useMemo(() => {
+    const lastUpdate = new Date(
+      siteConfig.customFields.lastResumeUpdate as string
+    );
+    const [, month, date, year] = lastUpdate.toDateString().split(" ");
+    return `${month} ${date} ${year}`;
+  }, [siteConfig.customFields.lastResumeUpdate]);
+
   return (
     <header className={clsx("hero hero--primary", styles.heroBanner)}>
       <div className="container">
         <h1 className="hero__title">Hi, I'm Charles!</h1>
         <p className="hero__subtitle">{siteConfig.tagline}</p>
-        {/* <div className={styles.buttons}>
-          <Link className="button button--secondary button--lg" to="/docs/hi">
-            Docusaurus Tutorial - 5min ⏱️
-          </Link>
-        </div> */}
+        <div className={styles.buttons}>
+          <a className="button button--secondary button--lg" href="/Resume.pdf">
+            Resume - Last Updated {lastResumeUpdate}
+          </a>
+        </div>
       </div>
     </header>
   );
